Make changes requested in https://github.com/zankich/devices/pull/2
- Combine test files into one test file
- Rename `NewLED` to `New`
- Export `Device` field in `LED`
